refactor: yield calculator refactor
- keep operation and position contexts apart
- use routing for component communication
- review folder structure to mimic DOM
