+++ libs/mstable/meta-vault/src/types.ts
@@ -5,7 +5,16 @@ import type {
   ChartType,
 } from './components/VaultPerformance/types';
 
+export type SupportedOperation = 'deposit' | 'mint' | 'withdraw' | 'redeem';
+
 export type MvGenerics = MakeGenerics<{
   Params: { mvid?: string };
-  Search: { chartType?: ChartType; chartTimeframe?: ChartTimeframe };
+  Search: {
+    chartType?: ChartType;
+    chartTimeframe?: ChartTimeframe;
+    input?: {
+      amount: number;
+      operation: SupportedOperation;
+    };
+  };
 }>;
