+++ libs/mstable/meta-vault/src/views/Metavault.tsx
@@ -12,7 +12,7 @@ import { Position } from '../components/Position';
 import { Strategy } from '../components/Strategy';
 import { VaultJumbo } from '../components/VaultJumbo';
 import { VaultPerformance } from '../components/VaultPerformance';
-import { MetavaultProvider, OperationsProvider } from '../state';
+import { MetavaultProvider } from '../state';
 
 import type { MvGenerics } from '../types';
 
@@ -30,29 +30,27 @@ export const Metavault = () => {
 
   return (
     <MetavaultProvider initialState={{ metavault }}>
-      <OperationsProvider>
-        <Stack direction="column">
-          <MstableBackground sx={(theme) => theme.mixins.paddings.jumbo}>
-            <VaultJumbo />
-          </MstableBackground>
-          <Box sx={(theme) => theme.mixins.paddings.page}>
-            <Grid container spacing={2}>
-              <Grid item xs={12} md={8} order={{ xs: 2, md: 1 }}>
-                <Stack direction="column" spacing={2}>
-                  <VaultPerformance />
-                  <Strategy />
-                </Stack>
-              </Grid>
-              <Grid item xs={12} md={4} order={{ xs: 1, md: 2 }}>
-                <Stack direction="column" spacing={2}>
-                  <Position />
-                  <Operations />
-                </Stack>
-              </Grid>
+      <Stack direction="column">
+        <MstableBackground sx={(theme) => theme.mixins.paddings.jumbo}>
+          <VaultJumbo />
+        </MstableBackground>
+        <Box sx={(theme) => theme.mixins.paddings.page}>
+          <Grid container spacing={2}>
+            <Grid item xs={12} md={8} order={{ xs: 2, md: 1 }}>
+              <Stack direction="column" spacing={2}>
+                <VaultPerformance />
+                <Strategy />
+              </Stack>
             </Grid>
-          </Box>
-        </Stack>
-      </OperationsProvider>
+            <Grid item xs={12} md={4} order={{ xs: 1, md: 2 }}>
+              <Stack direction="column" spacing={2}>
+                <Position />
+                <Operations />
+              </Stack>
+            </Grid>
+          </Grid>
+        </Box>
+      </Stack>
     </MetavaultProvider>
   );
 };
